Automatically track collectivite properties if present
